+++ .dotnet/src/Custom/VectorStores/StaticFileChunkingStrategy.cs
@@ -1,8 +1,8 @@
-using System;
-using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 
 namespace OpenAI.VectorStores;
 
+[Experimental("OPENAI001")]
 [CodeGenModel("StaticChunkingStrategyResponseParam")]
 public partial class StaticFileChunkingStrategy : FileChunkingStrategy
 {